.appveyor.yml: Retry Chocolatey installs and PhantomJS (#260)
* .appveyor.yml: Make phantomjs less flaky

It looks like the PhantomJS flakiness on TravisCI (see 3d11574) happens
on [AppVeyor] as well. This adds an [appveyor-retry] command to the `npm
test` portion of the AppVeyor test script, to workaround the
intermittent failures.

[AppVeyor]: https://ci.appveyor.com/project/stefanbuck/browser-extension/build/1.0.177
[appveyor-retry]: http://help.appveyor.com/discussions/suggestions/816-generic-wrapper-for-retry#comment_40504847

* .appveyor.yml: retry Chocolatey installs

(commit message copied from https://github.com/digitalcircuit/quassel/commit/19f50b66742c42f9139d34a61a3198981dfbe5ed)

Automatically retry up to 3 times on failure when installing
Chocolatey modules, using the 'appveyor-retry' script.  This can be
removed later if needed.

Sometimes Chocolatey returns a 404 for a package that exists.  This
happens a lot.  The official fix is to get the Business edition for
the more-stable private CDN, but since we're an open-source community
effort, sometimes a little kludgery goes a long way.

See the following:
* http://help.appveyor.com/discussions/suggestions/816-generic-wrapper-for-retry#comment_40579488
* https://github.com/appveyor/ci/issues/418