Test on Chrome/Firefox instead of PhantomJS, if possible (#286)
This uses [karma-detect-browsers] to determine if Chrome and Firefox are
present locally, and to test against them if so. If it can't find either
browser, it will use PhantomJS as before.

Additionally, both Travis and AppVeyor will have Chrome and Firefox
installed, so CI tests will use browsers as well. See the following
links for more information about these changes:

* https://ariya.io/2015/09/javascript-testing-with-latest-firefox-and-chrome-on-appveyor
* https://swizec.com/blog/how-to-run-javascript-tests-in-chrome-on-travis/swizec/6647
* https://docs.travis-ci.com/user/firefox/

[karma-detect-browsers]: https://www.npmjs.com/package/karma-detect-browsers

---

Note that I had to slightly tweak a test to get it to pass on both
Chrome and Firefox, but it's a test that has been flaky on PhantomJS in
the past, so I don't think it's an issue.